[REFACTOR] Modal컴포넌트 파일 하나로 통일
