+++ src/components/PetMeeting/PetMeeting.jsx
@@ -5,8 +5,7 @@ import FriendType from "./FriendType.jsx";
 import Scheduler from "./Scheduler.jsx";
 import { Box } from "@mui/material";
 import FindFriend from "./FindFriend.jsx";
-import PetConfigModal from "./PetConfigModal.jsx";
-import ActivityModal from "./ActivityModal.jsx";
+import { ActivityModal, PetConfigModal } from "./PetMeetingModals.jsx";
 
 const PetMeeting = () => {
     return (
