+++ src/components/PetMeeting/PetMeetingModals.jsx
@@ -1,11 +1,11 @@
 import React, { useContext, useEffect, useState } from "react";
+import { PetMeetingContext } from "../../context/PetMeetingContext.jsx";
 import { Box, Button, Fade, Modal, Typography } from "@mui/material";
 import Dog from "../../assets/images/PetMeeting/dog.svg";
 import SelectPetDropdown from "./SelectPetDropdown.jsx";
-import { PetMeetingContext } from "../../context/PetMeetingContext.jsx";
 
 const PetConfigModal = () => {
-    const { pet, setPet, open, drop, setDrop, setClose } = useContext(PetMeetingContext);
+    const { pet, setPet, openPetConfigModal, drop, setDrop, setClose } = useContext(PetMeetingContext);
     const [selectedPet, setSelectedPet] = useState(null);
     const [selectedActivity, setSelectedActivity] = useState(null);
 
@@ -17,7 +17,7 @@ const PetConfigModal = () => {
             setSelectedPet(null);
             setSelectedActivity(null);
         }
-    }, [pet, open]);
+    }, [pet, openPetConfigModal]);
 
     const petRegister = () => {
         setPet((prev) => ({ ...prev, name: selectedPet, activityStatus: selectedActivity }));
@@ -26,14 +26,14 @@ const PetConfigModal = () => {
 
     return (
         <Modal
-            open={open}
+            open={openPetConfigModal}
             onClose={setClose}
             disableScrollLock
             sx={{
                 zIndex: 10000,
             }}
         >
-            <Fade in={open} timeout={400}>
+            <Fade in={openPetConfigModal} timeout={400}>
                 <Box
                     sx={{
                         position: "absolute",
@@ -190,4 +190,183 @@ const PetConfigModal = () => {
     );
 };
 
-export default PetConfigModal;
+const ActivityModal = () => {
+    const { setPet, openActivityModal, setOpenActivityModal } = useContext(PetMeetingContext);
+
+    const handleRegisterClick = () => {
+        setPet((prev) => ({
+            ...prev,
+            activityStatus: "NONE",
+        }));
+        setOpenActivityModal(false);
+    };
+
+    return (
+        <Modal
+            open={openActivityModal}
+            onClose={() => setOpenActivityModal(false)}
+            disableScrollLock
+            sx={{
+                zIndex: 10000,
+            }}
+        >
+            <Fade in={openActivityModal} timeout={400}>
+                <Box
+                    sx={{
+                        display: "flex",
+                        flexDirection: "column",
+                        alignItems: "center",
+                        position: "absolute",
+                        top: "50%",
+                        left: "50%",
+                        transform: "translate(-50%, -50%)",
+                        width: 300,
+                        backgroundColor: "#FDF1E5",
+                        borderRadius: 2,
+                        p: 3,
+                        boxShadow: 24,
+                    }}
+                >
+                    <Box
+                        component="img"
+                        src={Dog}
+                        alt="dog"
+                        sx={{
+                            width: 60,
+                            height: 60,
+                            objectFit: "contain",
+                            marginBottom: "10px",
+                        }}
+                    ></Box>
+
+                    <Typography
+                        sx={{
+                            fontSize: "24px",
+                        }}
+                    >
+                        그만 놀까요?
+                    </Typography>
+                    <Box
+                        sx={{
+                            width: "100%",
+                            display: "flex",
+                            gap: "30px",
+                            justifyContent: "center",
+                            marginTop: "20px",
+                        }}
+                    >
+                        <Button
+                            sx={{
+                                backgroundColor: "#E9A260",
+                                color: "white",
+                                borderRadius: 5,
+                                padding: "7px 0",
+                                width: "100px",
+                            }}
+                            onClick={handleRegisterClick}
+                        >
+                            집가기
+                        </Button>
+                        <Button
+                            sx={{
+                                backgroundColor: "#F2DFCE",
+                                color: "black",
+                                borderRadius: 5,
+                                padding: "7px 0",
+                                width: "100px",
+                            }}
+                            onClick={() => setOpenActivityModal(false)}
+                        >
+                            취소
+                        </Button>
+                    </Box>
+                </Box>
+            </Fade>
+        </Modal>
+    );
+};
+
+/*const LocationSaveModal = () => {
+    <Modal
+        open={openActivityModal}
+        onClose={() => setOpenActivityModal(false)}
+        disableScrollLock
+        sx={{
+            zIndex: 10000,
+        }}
+    >
+        <Fade in={openActivityModal} timeout={400}>
+            <Box
+                sx={{
+                    display: "flex",
+                    flexDirection: "column",
+                    alignItems: "center",
+                    position: "absolute",
+                    top: "50%",
+                    left: "50%",
+                    transform: "translate(-50%, -50%)",
+                    width: 300,
+                    backgroundColor: "#FDF1E5",
+                    borderRadius: 2,
+                    p: 3,
+                    boxShadow: 24,
+                }}
+            >
+                <Box
+                    component="img"
+                    src={Dog}
+                    alt="dog"
+                    sx={{
+                        width: 60,
+                        height: 60,
+                        objectFit: "contain",
+                        marginBottom: "10px",
+                    }}
+                ></Box>
+
+                <Typography
+                    sx={{
+                        fontSize: "24px",
+                    }}
+                >
+                    그만 놀까요?
+                </Typography>
+                <Box
+                    sx={{
+                        width: "100%",
+                        display: "flex",
+                        gap: "30px",
+                        justifyContent: "center",
+                        marginTop: "20px",
+                    }}
+                >
+                    <Button
+                        sx={{
+                            backgroundColor: "#E9A260",
+                            color: "white",
+                            borderRadius: 5,
+                            padding: "7px 0",
+                            width: "100px",
+                        }}
+                        onClick={handleRegisterClick}
+                    >
+                        집가기
+                    </Button>
+                    <Button
+                        sx={{
+                            backgroundColor: "#F2DFCE",
+                            color: "black",
+                            borderRadius: 5,
+                            padding: "7px 0",
+                            width: "100px",
+                        }}
+                        onClick={() => setOpenActivityModal(false)}
+                    >
+                        취소
+                    </Button>
+                </Box>
+            </Box>
+        </Fade>
+    </Modal>;
+};*/
+export { ActivityModal, PetConfigModal };
