+++ src/context/PetMeetingContext.jsx
@@ -5,7 +5,7 @@ export const PetMeetingContext = createContext(null);
 
 export const PetMeetingProvider = ({ children }) => {
     const [pet, setPet] = useState(null);
-    const [open, setOpen] = useState(false);
+    const [openPetConfigModal, setOpenPetConfigModal] = useState(false);
     const [openActivityModal, setOpenActivityModal] = useState(false);
     const [drop, setDrop] = useState(false);
     const [friendType, setFriendType] = useState("산책친구들");
@@ -13,26 +13,26 @@ export const PetMeetingProvider = ({ children }) => {
 
     const setClose = () => {
         setDrop(false);
-        setOpen(false);
+        setOpenPetConfigModal(false);
     };
 
     const value = useMemo(
         () => ({
             pet,
-            open,
+            openPetConfigModal,
             drop,
             friendType,
             view,
             openActivityModal,
             setPet,
-            setOpen,
+            setOpenPetConfigModal,
             setDrop,
             setFriendType,
             setView,
             setOpenActivityModal,
             setClose,
         }),
-        [pet, open, drop, friendType, view, openActivityModal]
+        [pet, openPetConfigModal, drop, friendType, view, openActivityModal]
     );
 
     useEffect(() => {
